Adicionando classes DTOs

+++ src/main/java/com/example/bank_blue/repository/ClienteRepository.java
@@ -1,6 +1,6 @@
 package com.example.bank_blue.repository;
 
-import com.example.bank_blue.model.clienteModel.Cliente;
+import com.example.bank_blue.model.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 @Repository
 public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
-    Optional<Cliente> findClienteById(Long id);
+    Optional<Cliente> findClienteById(Integer id);
     Optional<Cliente> findClienteByCpfCnpj(String cpfCnpj);
 
     boolean existsByCpfCnpj(String cpfCnpj);
