+++ src/main/java/com/example/bank_blue/repository/EnderecoRepository.java
@@ -1,6 +1,6 @@
 package com.example.bank_blue.repository;
 
-import com.example.bank_blue.model.endereco.Endereco;
+import com.example.bank_blue.model.Endereco;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,6 +8,7 @@ import java.util.Optional;
 
 @Repository
 public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {
-    Optional<Endereco> findEnderecoById(long id);
-    boolean existsById(long id);
+
+    Optional<Endereco> findEnderecoById(Integer id);
+    boolean existsById(Integer id);
 }
